requirements.txt
# web api
gevent
gevent-psycopg2
futures
gunicorn
falcon
falcon-cors

# necessary for https and certs
requests
urllib3
pyopenssl
ndg-httpsclient
pyasn1

dateparser
fuzzywuzzy
python-Levenshtein

geoalchemy2
geojson
shapely
pytz
lxml
xmltodict

iso8601
Jinja2
numpy
ujson

dogpile.cache

XlsxWriter

click

# harvester deps
boto3
grequests
decorator